refactor: Simplify imports from src/Config/Themes
