+++ src/Config/Themes/index.ts
@@ -1,0 +1,2 @@
+export * from './ITSThemeCollection';
+export * from './MinimalThemeCollection';
